Safety commit before attempting fix
In the process of porting over Johannes' camera and death/respawn system to my C++ character class. Currently have a problem where I have both the old Camera and SpringArm despite having changed them in the code to new FollowCamera and CameraBoom. Will attempt to delete the old BP_Player and make a new one and see if that fixes it.