Semi fixed update
I think the movement works more like it should now, but there're issues with the camera, possibly because the camera offset system was set up to  deal with the default thirdperson class code and not the code I made for my player class. Will attempt to get to the bottom of that next.