Game Over screen mostly functional
Game Over / Death screen with Respawn and Main Menu buttons is up and running; I don't have a main menu yet, but I'll get to that after merging and updating.

Some more features that would be nice: Add a delay so that the whole getting eaten by a fish animation plays first before the Game Over screen pops up.

Also try to clean up and consolidate the code as much as possible to prevent merging conflicts; Ideally everything should be done in the Player Controller and Widget Blueprints since those aren't really being touched atm, but right now the whole load new level on death functionality is being done in the player character blueprint, so if for example my Respawn button from the Game Over screen is going to make use of that, I have to figure out a way in which it will create as little merging conflicts as possible.

Most important now before proceeeding is just to get this stuff merged back into the Main Niklas branch and then updating that from main, and doing that in a way in a way that doesn't result in errors.