End-of-tutorial commit
Reached the end of the tutorial. Problems to be fixed: In the video he sets up the options menu to toggle by pressing escape. I have that behavior for my Pause Menu, so I want to access the Options Menu by pressing its respective button in either the Main Menu or Pause Menu. I also think I should add a "Back" button to the Options menu so that you can click that to return to whichever menu you were in when you entered the Options Menu.

Another problem I've found is that when you apply audio level changes in the options menu, that works, UNTIL you set a slider to 0; After that, even if you put the slider up again and press apply, the sound remains off.